perf(web): lazy-load AgentPanel + analytics charts, drop redundant force-dynamic
Feature-preserving performance tuning (no page/route/capability removed; route
count stays 152). All changes verified with tsc (clean).

- Code-split the floating AI AgentPanel via next/dynamic ssr:false in
  dashboard-shell — it starts closed and drags in framer-motion + voice capture,
  so its chunk no longer ships in every dashboard page's first load.
- Split recharts out of the /analytics first-load: moved the revenue BarChart +
  quote PieChart into analytics-charts.tsx, dynamic-imported with skeleton
  fallbacks so the KPI shell paints before the chart lib loads.
- Removed the redundant `export const dynamic = "force-dynamic"` from 25
  (dashboard) pages — the segment is already dynamic via cookies()/headers() in
  its layout, so the directive was a no-op and the exact anti-pattern the perf
  notes warn against. Kept it on admin + public token pages (no shared dynamic
  layout / must never be statically cached).

Deferred (need the networked preview to do safely):
- Self-hosting Google fonts (needs the .woff2 files downloaded).
- Slimming select("*") getters — the high-payoff ones (getCustomers/getProducts)
  are shared getters used by pickers/forms; slimming risks consumers and can't be
  validated without a preview build.
