+++ src/app/(dashboard)/contracts/[id]/page.tsx
@@ -5,8 +5,6 @@ import { canEdit as roleCanEdit, type Role } from "@/lib/permissions";
 import { getContract, renderContractHtml } from "@/lib/actions/contracts";
 import { ContractDetailClient } from "@/components/contracts/contract-detail-client";
 
-export const dynamic = "force-dynamic";
-
 export default async function ContractDetailPage({ params }: { params: Promise<{ id: string }> }) {
   const { id } = await params;
   const supabase = await createClient();
