+++ src/app/(dashboard)/prospects/[id]/page.tsx
@@ -4,8 +4,6 @@ import { getActiveBizId } from "@/lib/active-business";
 import { getProspect, listProspectActivities } from "@/lib/actions/prospects";
 import { ProspectDetail } from "@/components/prospects/prospect-detail";
 
-export const dynamic = "force-dynamic";
-
 export default async function ProspectDetailPage({ params }: { params: Promise<{ id: string }> }) {
   const { id } = await params;
   const supabase = await createClient();
