+++ src/app/(dashboard)/seo/content/[id]/page.tsx
@@ -5,8 +5,6 @@ import { getContentPiece } from "@/lib/actions/seo-pipeline";
 import { listConnections } from "@/lib/actions/seo-connections";
 import { SeoContentDetail } from "@/components/seo/seo-content-detail";
 
-export const dynamic = "force-dynamic";
-
 export default async function SeoContentDetailPage({ params }: { params: Promise<{ id: string }> }) {
   const { id } = await params;
   const supabase = await createClient();
